Add method to search pull requests
Add methodf to search pull requests given a search string.
