+++ bruno/GitHub/Pull requests/Get pull request review comments.bru
@@ -1,7 +1,7 @@
 meta {
   name: Get pull request review comments
   type: http
-  seq: 2
+  seq: 4
 }
 
 get {
