+++ bruno/GitHub/Pull requests/Get pull requests in a repository.bru
@@ -1,7 +1,7 @@
 meta {
-  name: Get pull request
+  name: Get pull requests in a repository
   type: http
-  seq: 3
+  seq: 1
 }
 
 get {
