+++ bruno/GitHub/Pull requests/List issue comments.bru
@@ -1,7 +1,7 @@
 meta {
   name: List issue comments
   type: http
-  seq: 1
+  seq: 3
 }
 
 get {
