+++ bruno/GitHub/Pull requests/Search pull requests.bru
@@ -1,0 +1,43 @@
+meta {
+  name: Search pull requests
+  type: http
+  seq: 2
+}
+
+get {
+  url: https://api.github.com/search/issues?q&per_page=100
+  body: none
+  auth: inherit
+}
+
+params:query {
+  q: 
+  per_page: 100
+}
+
+settings {
+  encodeUrl: false
+  timeout: 0
+}
+
+docs {
+  ## Search query
+  
+  Here is an example you can use as reference search query (parameter `q`) to search pull requests:
+  
+  ```none
+  <KEYWORD> is:pr created:2025-07-01..2025-07-15 -repo:<ORGANIZATION>/<REPO>
+  ```
+  
+  Where:
+  
+  * `<KEYWORD>` is a search keyword.
+  * `is:pr` specifies that result must include only pull requests.
+  * `created:…` limits the result including only the pull requests created in the given interval.
+  * `-repo:…` excludes the pull requests of the specified repository.
+  
+  ## Additional documentation
+  
+  * API documentation: [Searching issues and pull requests](https://docs.github.com/en/rest/search/search?apiVersion=2022-11-28#search-issues-and-pull-requests)
+  * Search queries: [Searching issues and pull requests](https://docs.github.com/en/search-github/searching-on-github/searching-issues-and-pull-requests)
+}
