+++ bruno/GitHub/collection.bru
@@ -1,5 +1,6 @@
 headers {
   X-GitHub-Api-Version: 2022-11-28
+  Accept: application/vnd.github+json
 }
 
 auth {
